Fix Ras Safaniya runway 14 texture z-fighting issues (#11)
The the northern segment of the runway was 1 foot lower than the
rest of the runway. This caused the runway model to clash with the
surrounding terrain, creating a very distracting texture glitch when
flying on the runway.

This changes some of the coastal tile elevations north of the base
form 0ft MSL to 3ft MSL to match the rest of the base. That seems to
fully flatten the airbase tiles and fix the z-fighting issues.